Tight mode, but in JS!
This prevents browsers from spending bandwidth preloading loupe
thumbnails before the contact sheet is loaded.

This is not an ideal solution, as it wastes a full round-trip doing
nothing. HTTP2-3 content prioritization is supposed to solve this
problem without incurring in additional latency, but as of July 2025,
browser -- and especially server -- implementations are not reliable
enough to be able to safely depend on them for a hobby project.