Preload enough resolution for rotated full-screen
By using the 'sizes' attribute of `<source>` it's possible to tell the
browser how wide (in CSS units) the image should be considered to be for
the purposes of choosing a variant in the srcset.

We can use this so that the browser not only gets enough resolution for
the current viewport, but also to account for rotations and, with some
JS, for full-screen.